refactor sdb,c2kd
