+++ README.md
@@ -352,6 +352,8 @@ Please follow the instructions in [QuickStart.md](docs/QuickStart.md) for the ba
   year={2022}
 }
 ```
+
+**Scripts can be found in [examples/distillation/sdb](examples/distillation/sdb)**
 </details>
 
 <details>
@@ -403,6 +405,8 @@ url = "https://www.jcad.cn/article/doi/10.3724/SP.J.1089.2022.19723",
 author = "赵雅","冯尊磊","王慧琼","宋明黎",keywords = "唇语识别","知识蒸馏","跨模态",
 }
 ```
+
+**Scripts can be found in [examples/distillation/c2kd](examples/distillation/c2kd)**
 </details>
