--- examples/C2KD/README.md
@@ -1,11 +1,0 @@
-# Context Correlation Distillation for Lip Reading
-
-Please Note that this code does not provide the entire training process for the dataset's extracted features missing.
-
-This code only provide the main part, like losses and models.
-
-The processing methods for LSR datasets should be differently achieved by yourself.
-
-The original model files is available in C2KD/model/ and the loss files is in C2KD/utils/losses.py.
-
-The code under the KAE framework is available in kamal.slim.distillation.C2KD.
