+++ examples/distillation/c2kd/utils/12_6_2048.yaml
