+++ examples/distillation/sdb/train_sdb_teacher.py
@@ -1,4 +1,3 @@
-# train a baseline model from scratch
 import sys
 sys.path.append("../../../")
 
@@ -16,7 +15,8 @@ import kamal
 from kamal.vision import sync_transforms as sT
 from torch.utils.tensorboard import SummaryWriter
 from kamal import vision, engine, callbacks
-from kamal.amalgamation import AdversTeacher,AdversTEvaluator,KD_SDB_Stuednt
+from kamal.distillation.sdb.safe_distillation_box import AdversTeacher,AdversTEvaluator,KD_SDB_Stuednt
+from kamal.distillation.sdb.sdb_task import SDBTask,KD_SDB_Task
 
 
 # random seed
@@ -34,9 +34,9 @@ parser = argparse.ArgumentParser()
 parser.add_argument('--noise_path', default='./CIFAR10/', type=str)
 parser.add_argument('--dataset', default='cifar10', type=str)
 parser.add_argument('--resume', default=None, type=str)
-parser.add_argument('--stu_model_name', default='resnet18', type=str)
-parser.add_argument('--teacher_model', default='resnet18', type=str)
-parser.add_argument('--teacher_resume', default='./checkpoints/cifar10_SDBtch_acc_0.8652.pth', type=str)
+parser.add_argument('--model_name', default='resnet18', type=str)
+parser.add_argument('--adversarial_model', default='resnet18', type=str)
+parser.add_argument('--adversarial_resume', default='./CIFAR10/cifar10_wrn18.pth', type=str)
 
 parser.add_argument('--learning_rate', default=0.1, type=float)
 parser.add_argument('--batch_size', default=128, type=int)
@@ -46,7 +46,6 @@ parser.add_argument('--temperature', default=4, type=float)
 parser.add_argument('--weight', default=1, type=float)
 parser.add_argument('--lamb', default=0.9, type=float)
 parser.add_argument('--eta', default=0.01, type=float)
-parser.add_argument('--alpha', default=1, type=float)
 parser.add_argument('--cuda', default=True, type=bool)
 args = parser.parse_args()
 
@@ -64,7 +63,8 @@ def generate_noise(params):
         noisy_img = np.random.randint(0, 255, size=(64, 64, 3))
         return data_transformer(noisy_img / 255.0).float()
 
-def main(args): 
+
+def main(args):
     # Dataset
     train_dst = vision.datasets.torchvision_datasets.CIFAR10( 
         'data/data-cifar10', train=True, download=True, transform=sT.Compose([
@@ -81,58 +81,51 @@ def main(args):
     train_loader = torch.utils.data.DataLoader( train_dst, batch_size=args.batch_size, shuffle=True, num_workers=args.num_workers )
     val_loader = torch.utils.data.DataLoader( val_dst, batch_size=args.batch_size, num_workers=args.num_workers )
 
-     # Model 
+    # SDB Model 
     if args.dataset == 'cifar10':
         num_class = 10
 
     print('Number of class: ' + str(num_class))
-    print('Create Student Model --- ' + args.stu_model_name)
+    print('Create Model --- ' + args.model_name)
 
-    if args.stu_model_name == 'resnet18':
-        model = vision.models.classification.resnet18(num_classes=num_class)
+    if args.model_name == 'resnet18':
+        model = vision.models.classification.cifar.wrn.wrn_40_2(num_classes=num_class)
     else:
         model = None
-        print('Not support for model ' + str(args.stu_model_name))
+        print('Not support for model ' + str(args.model_name))
         exit()
 
-    # Teacher Model
-    if args.teacher_model == 'resnet18':
-        teacher_model = vision.models.classification.cifar.wrn.wrn_40_2(num_classes=num_class)
-    else:
-        teacher_model = None
-        print('Not support for model ' + str(args.teacher_model))
-        exit()
+    # Adversarial model
+    print('Create Adversarial Model --- ' + args.adversarial_model)
+
+    if args.adversarial_model == 'resnet18':
+        adversarial_model = vision.models.classification.cifar.wrn.wrn_40_2(num_classes=num_class)
+        random_model = vision.models.classification.cifar.wrn.wrn_40_2(num_classes=num_class)
 
     if args.cuda:
         model = model.cuda()
-        teacher_model = teacher_model.cuda()
+        adversarial_model = adversarial_model.cuda()
+        random_model = random_model.cuda()
 
     # checkpoint
     if args.resume:
-        print('- Load checkpoint model from {}'.format(args.resume))
+        print("Load checkpoint from {}".format(args.resume))
         checkpoint = torch.load(args.resume)
         model.load_state_dict(checkpoint)
     else:
-        print('- Train from scratch ')
+        print('- Train SDB teacher model from scratch ')
 
-    # load teacher model
-    if args.teacher_resume:
-        teacher_resume = args.teacher_resume
-        print('------ Teacher Resume from {}'.format(teacher_resume))
+    # load trained Adversarial model
+    if(args.adversarial_resume != "None"):
+        print("Load Trained adversarial model")
+        checkpoint = torch.load(args.adversarial_resume)
+        adversarial_model.load_state_dict(checkpoint)
     else:
-        print('Please Load a Trained Teacher!')
-    checkpoint = torch.load(teacher_resume)
-    teacher_model.load_state_dict(checkpoint)
+        print("Please Use a Training advers_teacher")
 
     # Optimizer
-    if args.stu_model_name == 'CNN' or args.stu_model_name == 'mlp':
-        optimizer = Adam(model.parameters(), lr=args.learning_rate)
-        print('Optimizer: Adam')
-    else:
-        optimizer = SGD(model.parameters(), lr=args.learning_rate, momentum=0.9, weight_decay=5e-4)
-        print('Optimizer: SGD')
+    optimizer = SGD(model.parameters(), lr=args.learning_rate, momentum=0.9, weight_decay=5e-4)
     TOTAL_ITERS= len(train_loader) * args.num_epochs
-    device = torch.device( 'cuda' if torch.cuda.is_available() else 'cpu' )
     sched = torch.optim.lr_scheduler.CosineAnnealingLR( optimizer, T_max=TOTAL_ITERS )
 
     # KAE Part
@@ -150,33 +143,37 @@ def main(args):
 
     # prepare evaluator
     metric = kamal.tasks.StandardMetrics.classification()
-    evaluator = engine.evaluator.BasicEvaluator(dataloader=val_loader, metric=metric, progress=False)
+    evaluator = AdversTEvaluator(dataloader=val_loader, metric=metric, progress=False, params=args,noise = noise_data,)
 
     # prepare trainer
-    task = kamal.tasks.KD_SDB_Task(name='KD_SDB_Teacher',
-            loss_fn_kd=nn.KLDivLoss(reduction='batchmean'),
-            loss_fn_ce=nn.CrossEntropyLoss(),
-            scaling=1.0, 
-            pred_fn=lambda x: x.max(1)[1], 
-            attach_to=None
-    )
-    trainer = KD_SDB_Stuednt( 
-        logger=kamal.utils.logger.get_logger('cifar10'), 
+    task = SDBTask(name='SDB_Teacher_train', 
+                    loss_fn_ce=nn.CrossEntropyLoss(), 
+                    loss_fn_kd=nn.KLDivLoss(reduction='batchmean'), 
+                    loss_fn_mse=nn.MSELoss(reduction='mean'),
+                    scaling=1.0, 
+                    pred_fn=lambda x: x.max(1)[1], 
+                    attach_to=None)
+
+    trainer = AdversTeacher( 
+        logger=kamal.utils.logger.get_logger('cifar10-sdb'), 
         tb_writer=SummaryWriter( log_dir='run/cifar10-%s'%( time.asctime().replace( ' ', '_' ) ) ) 
     )
-    trainer.setup( student=model, 
-                teacher=teacher_model,
-                task=task, 
-                dataloader=train_loader,
-                optimizer=optimizer,
-                params=args,
-                noise=noise_data,
-                device=device)
+
+    trainer.setup( model=model, 
+                   advers_model=adversarial_model,
+                   random_model=random_model,
+                   task=task, 
+                   dataloader=train_loader,
+                   optimizer=optimizer,
+                   device=device,
+                   params=args,
+                   noise=noise_data,
+                )
 
     # add callbacks
     trainer.add_callback( 
         engine.DefaultEvents.AFTER_EPOCH, 
-        callbacks=callbacks.EvalAndCkpt(model=model, evaluator=evaluator, metric_name='acc', ckpt_prefix='cifar10_SDB_KD_student') )
+        callbacks=callbacks.EvalAndCkpt(model=model, evaluator=evaluator, metric_name='acc', ckpt_prefix='cifar10_SDBtch') )
     trainer.add_callback(
         engine.DefaultEvents.AFTER_STEP,
         callbacks=callbacks.LRSchedulerCallback(schedulers=[sched]))
