+++ kamal/core/tasks/__init__.py
@@ -1,3 +1,3 @@
-from .task import StandardTask, StandardMetrics, GeneralTask, Task, TaskCompose, SDBTask, KD_SDB_Task
+from .task import StandardTask, StandardMetrics, GeneralTask, Task, TaskCompose
 from . import loss
